+++ src/fireset/server.py
@@ -1,94 +1,17 @@
-import functools
-import inspect
-import typing as T
-from dataclasses import dataclass
 from datetime import datetime
 import xml.etree.ElementTree as ET
 
 import logfire
-from starlette.exceptions import HTTPException
 from starlette.applications import Starlette
 from starlette.routing import Route
 from starlette.requests import Request
 from starlette.responses import Response, PlainTextResponse
 from starlette.middleware import Middleware
 from starlette.middleware.authentication import AuthenticationMiddleware
-from starlette_auth_toolkit.base.backends import BaseBasicAuth
-from starlette_auth_toolkit.cryptography import PBKDF2Hasher
 
 from . import logger, settings
-from .database import get_db_vcard, list_db_vcards, get_contact_from_email
-
-
-_P = T.ParamSpec("_P")
-
-# Password hasher
-hasher = PBKDF2Hasher()
-
-
-def authenticated(func: T.Callable[_P, T.Any]) -> T.Callable[_P, T.Any]:
-    """Decorator that checks whether the request is authenticated,
-    and ig the user has the right to access the resource
-
-    """
-    sig = inspect.signature(func)
-    for idx, parameter in enumerate(sig.parameters.values()):
-        if parameter.name == "request" or parameter.name == "websocket":
-            break
-    else:
-        raise Exception(f'No "request" or "websocket" argument on function "{func}"')
-
-    # Handle async request/response functions.
-    @functools.wraps(func)
-    async def async_wrapper(*args, **kwargs) -> T.Any:
-        request = kwargs.get("request", args[idx] if idx < len(args) else None)
-        assert isinstance(request, Request)
-        user: BookUser = request.user
-        user_id = int(request.path_params.get("user_id", -1))
-
-        if "authenticated" not in request.auth.scopes or user.id != user_id:
-            raise HTTPException(status_code=403)
-
-        return await func(*args, **kwargs)
-
-    return async_wrapper
-
-
-@dataclass
-class BookUser:
-    id: int
-    email: str
-    password: str
-    is_authenticated: bool
-
-    @classmethod
-    def from_db(cls, email: str) -> "BookUser":
-        db_contact = get_contact_from_email(email)
-        if db_contact is None:
-            return None
-
-        if db_contact.hashed_password is None:
-            return None
-
-        res = cls(
-            id=db_contact.id,
-            email=email,
-            password=db_contact.hashed_password,
-            is_authenticated=False,
-        )
-
-        return res
-
-
-# Authentication backend
-class BasicAuth(BaseBasicAuth):
-    async def find_user(self, username: str):
-        u = BookUser.from_db(email=username)
-        return u
-
-    async def verify_password(self, user: BookUser, password: str):
-        user.is_authenticated = await hasher.verify(password, user.password)
-        return user.is_authenticated
+from .database import get_db_vcard, list_db_vcards
+from .authentication import BasicAuth, authenticated, BookUser
 
 
 @authenticated
@@ -125,7 +48,7 @@ async def handle_options(request: Request):
     return Response(
         headers={
             "Allow": "OPTIONS, GET, PUT, DELETE, PROPFIND, REPORT",
-            "DAV": "1, 3, addressbook",
+            "DAV": "1, 2, addressbook",
             "Content-Length": "0",
         },
     )
@@ -133,10 +56,14 @@ async def handle_options(request: Request):
 
 @authenticated
 async def handle_propfind(request: Request):
+    user: BookUser = request.user
+    user_id = int(request.path_params.get("user_id", -1))
+    assert user.id == user_id
+
     depth = request.headers.get("Depth", "0")
     content_length = int(request.headers.get("Content-Length", 0))
     request_body = await request.body()
-    # request_body = self.rfile.read(content_length)
+
     if content_length > 0:
         try:
             # Parse the XML request
@@ -153,9 +80,9 @@ async def handle_propfind(request: Request):
     if depth == "0":
         # Listing collections
         response = '<?xml version="1.0" encoding="UTF-8"?>\n'
-        response += '<D:multistatus xmlns:D="DAV:">\n'
+        response += '<D:multistatus xmlns:D="DAV:" xmlns:C="urn:ietf:params:xml:ns:carddav">\n'
         response += "  <D:response>\n"
-        response += "    <D:href>/</D:href>\n"
+        response += f"    <D:href>/users/{user_id}</D:href>\n"
         response += "    <D:propstat>\n"
         response += "      <D:prop>\n"
         response += "        <D:resourcetype>\n"
@@ -223,9 +150,9 @@ async def handle_wall_known(request: Request):
 app = Starlette(
     routes=[
         Route("/.well-known/carddav", handle_wall_known, methods=["GET"]),
-        Route("/{card_id}", handle_get, methods=["GET"]),
         Route("/users/{user_id}", handle_options, methods=["OPTIONS"]),
-        Route("/", handle_propfind, methods=["PROPFIND"]),
+        Route("/users/{user_id}", handle_propfind, methods=["PROPFIND"]),
+        Route("/{card_id}", handle_get, methods=["GET"]),
     ],
     middleware=[
         Middleware(